chore: Update RewardsStatistic import in DetailsSubInfo
